Fixing bug wherein the default was assumed to be variables.
Also making it where we now treat anything (ideally, comments just inside the CREATE statement) to be descriptions for the package/type.

JSDoc assumes the first comment will come **before** the first statement, but PL/SQL actively works against that: line offsets are based off the line with CREATE, not the first line in a "file", because to Oracle it is just a statement.