Fixing number matching so we match numbers starting with '.'.
This requires two things:

1. Move "number" rule before "operator".
2. Drop the unary sign from the number rule.

The problem is because '.' is also an operator, having "number" after that means we never match ".1" as a valid value,
which PL/SQL does.

If we simply move "number" before "operator", then we lose the ability to properly handle binary expressions like "1+1".

Since we are already handling unary sign detection as a separate token, removing the leading sign matching from the number token resolves both cases.